Moving Quicklinks code into its own file
