+++ wp-content/themes/wp-shield/functions.php
@@ -18,6 +18,7 @@
 require_once( 'library/wp-shield/customizer/logo-uploader.php' );
 require_once( 'library/wp-shield/customizer/customizr-tweaks.php' );
 require_once( 'library/wp-shield/customizer/register-menus.php' );
+require_once( 'library/wp-shield/customizer/quicklinks.php' );
 require_once( 'library/wp-shield/customizer/theme-design-switch.php' );
 require_once( 'library/wp-shield/customizer/footer.php' );
 require_once( 'library/wp-shield/banners/metaboxes-page-banners.php' );
